Moved addToCart function to App
