+++ cgWebShop/frontend/src/App.jsx
@@ -7,6 +7,7 @@ import Login from './Pages/Login'
 import Register from './Pages/Register'
 import Account from './Pages/Account'
 import Order from './Pages/Order'
+import ProductInfo from './Pages/ProductInfo'
 
 export const Context = createContext();
 
@@ -51,8 +52,40 @@ function App() {
 
   const user = localStorage.getItem("user") ? JSON.parse(localStorage.getItem("user")) : null;
 
+  const addToCart = async (product) => {
+    const response = await fetch(`/api/Cart/Item?userId=${user.id}&quantity=1`, {
+      method: 'POST',
+      headers: {
+        'Content-Type': 'application/json'
+      },
+      body: JSON.stringify({
+        id: product.id,
+        category: {
+          id: product.category.id,
+          name: product.category.name,
+          products: [
+          ]
+        },
+        name: product.name,
+        inStock: product.inStock,
+        price: product.price,
+        img: product.img
+      })
+    })
+
+    if(response.ok){
+      const cartResp = await fetch(`/api/Cart?userId=${user.id}`);
+
+      if(cartResp.ok){
+        const result = await cartResp.json();
+        setCart(result.cartItems)
+        setCartCounter(result.cartItems.length)
+      };
+    };
+  }
+
   return (
-    <Context.Provider value={{ cartCounter, setCartCounter, setShowAcc, showAcc, products, isAuthenticated, setIsAuthenticated, filteredProducts, setFilteredProducts, searchValue, setSearchValue, setIsAdmin, isAdmin, cart, setCart, user, categories, setCategoriesRefresh, categoriesRefresh, productsRefresh, setProductsRefresh }}>
+    <Context.Provider value={{ addToCart, cartCounter, setCartCounter, setShowAcc, showAcc, products, isAuthenticated, setIsAuthenticated, filteredProducts, setFilteredProducts, searchValue, setSearchValue, setIsAdmin, isAdmin, cart, setCart, user, categories, setCategoriesRefresh, categoriesRefresh, productsRefresh, setProductsRefresh }}>
       <BrowserRouter>
         <Header />
         <Routes>
@@ -61,6 +94,7 @@ function App() {
           <Route path='/register' element={<Register />}/>
           <Route path='/account' element={<Account />}/>
           <Route path='/order' element={<Order />} />
+          <Route path='/products/:productName' element={<ProductInfo products={products} addToCart={addToCart}/>} />
         </Routes>
       </BrowserRouter>
     </Context.Provider>
